backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.10
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# SEO and Analytics
pytrends==4.9.2
beautifulsoup4==4.12.3
requests==2.31.0

# AI and Content Generation
openai==1.12.0

# Caching and Performance
redis==5.0.1
boto3==1.34.34

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-slugify==8.0.4
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.3
schedule==1.2.1

# Additional dependencies
tweepy==4.14.0
instaloader==4.10.2

# New dependencies
python-dateutil==2.8.2
gunicorn==21.2.0 